+++ packages/store-sync/src/blockEventsToStorage.ts
@@ -7,17 +7,15 @@ import {
   schemaIndexToAbiType,
 } from "@latticexyz/protocol-parser";
 import { GroupLogsByBlockNumberResult, GetLogsResult } from "@latticexyz/block-logs-stream";
-import { StoreEventsAbi, StoreConfig, ExpandTablesConfig } from "@latticexyz/store";
+import { StoreEventsAbi, StoreConfig } from "@latticexyz/store";
 import { TableId } from "@latticexyz/common";
 import { Hex, decodeAbiParameters, parseAbiParameters } from "viem";
 import { debug } from "./debug";
 // TODO: move these type helpers into store?
 import { Key, Value } from "@latticexyz/store-cache";
+import { isDefined } from "@latticexyz/common/utils";
 
 // TODO: change table schema/metadata APIs once we get both schema and field names in the same event
-// TODO: support passing in a MUD config to get typed tables, values, etc.
-// TODO: consider if we should continue to group storage operations by block, allowing atomic sets (db txs) and collapsing operations per table+key
-//       or potentially have a start/end transaction and pass in the tx to each storage operation
 
 // TODO: export these from store or world
 export const schemaTableId = new TableId("mudstore", "schema");
@@ -39,46 +37,56 @@ export type StoredTableMetadata = {
   valueNames: readonly string[];
 };
 
-export type SetRecordOptions<
-  TConfig extends StoreConfig,
-  TTableName extends string = string & keyof ExpandTablesConfig<TConfig["tables"]>
-> = {
+export type SetRecordOperation<TConfig extends StoreConfig> = {
+  type: "SetRecord";
+  address: string;
   namespace: string;
-  name: TTableName;
-  keyTuple: Key<TConfig, TTableName>;
-  record: Value<TConfig, TTableName>;
-};
+} & {
+  [TTable in keyof TConfig["tables"]]: {
+    name: TTable;
+    keyTuple: Key<TConfig, TTable>;
+    record: Value<TConfig, TTable>;
+  };
+}[keyof TConfig["tables"]];
 
-export type SetFieldOptions<
-  TConfig extends StoreConfig,
-  TTableName extends string = string & keyof ExpandTablesConfig<TConfig["tables"]>,
-  TValueName extends string = string & keyof Value<TConfig, TTableName>
-> = {
+export type SetFieldOperation<TConfig extends StoreConfig> = {
+  type: "SetField";
+  address: string;
   namespace: string;
-  name: string;
-  keyTuple: Key<TConfig, TTableName>;
-  // TODO: standardize on calling these "fields" or "values" or maybe "columns"
-  valueName: TValueName;
-  value: Value<TConfig, TTableName>[TValueName];
-};
+} & {
+  [TTable in keyof TConfig["tables"]]: {
+    name: TTable;
+    keyTuple: Key<TConfig, TTable>;
+  } & {
+    [TValue in keyof Value<TConfig, TTable>]: {
+      // TODO: standardize on calling these "fields" or "values" or maybe "columns"
+      valueName: TValue;
+      value: Value<TConfig, TTable>[TValue];
+    };
+  }[keyof Value<TConfig, TTable>];
+}[keyof TConfig["tables"]];
 
-export type DeleteRecordOptions<
-  TConfig extends StoreConfig,
-  TTableName extends string = string & keyof ExpandTablesConfig<TConfig["tables"]>
-> = {
+export type DeleteRecordOperation<TConfig extends StoreConfig> = {
+  type: "DeleteRecord";
+  address: string;
   namespace: string;
-  name: string;
-  keyTuple: Key<TConfig, TTableName>;
-};
+} & {
+  [TTable in keyof TConfig["tables"]]: {
+    name: TTable;
+    keyTuple: Key<TConfig, TTable>;
+  };
+}[keyof TConfig["tables"]];
 
-export type BlockEventsToStorageOptions<TConfig extends StoreConfig = StoreConfig> = {
+export type StorageOperation<TConfig extends StoreConfig> =
+  | SetFieldOperation<TConfig>
+  | SetRecordOperation<TConfig>
+  | DeleteRecordOperation<TConfig>;
+
+export type BlockEventsToStorageOptions = {
   registerTableSchema: (data: StoredTableSchema) => Promise<void>;
   registerTableMetadata: (data: StoredTableMetadata) => Promise<void>;
   getTableSchema: (opts: Pick<StoredTableSchema, "namespace" | "name">) => Promise<StoredTableSchema | undefined>;
   getTableMetadata: (opts: Pick<StoredTableMetadata, "namespace" | "name">) => Promise<StoredTableMetadata | undefined>;
-  setRecord: (opts: SetRecordOptions<TConfig>) => Promise<void>;
-  setField: (opts: SetFieldOptions<TConfig>) => Promise<void>;
-  deleteRecord: (opts: DeleteRecordOptions<TConfig>) => Promise<void>;
 };
 
 export function blockEventsToStorage<TConfig extends StoreConfig = StoreConfig>({
@@ -86,10 +94,11 @@ export function blockEventsToStorage<TConfig extends StoreConfig = StoreConfig>(
   registerTableSchema,
   getTableMetadata,
   getTableSchema,
-  setRecord,
-  setField,
-  deleteRecord,
-}: BlockEventsToStorageOptions<TConfig>): (block: BlockEvents) => Promise<void> {
+}: BlockEventsToStorageOptions): (block: BlockEvents) => Promise<{
+  blockNumber: BlockEvents["blockNumber"];
+  blockHash: BlockEvents["blockHash"];
+  operations: StorageOperation<TConfig>[];
+}> {
   return async (block) => {
     // Find and register all new table schemas
     // Store schemas are immutable, so we can parallelize this
@@ -138,54 +147,89 @@ export function blockEventsToStorage<TConfig extends StoreConfig = StoreConfig>(
       })
     );
 
-    // Because storage operations are atomic, we need to do this serially
-    // TODO: make this smarter, maybe parallelize by table?
-    for (const log of block.logs) {
-      const tableId = TableId.fromHex(log.args.table);
-      const [tableSchema, tableMetadata] = await Promise.all([getTableSchema(tableId), getTableMetadata(tableId)]);
-      if (!tableSchema) {
-        debug("no table schema found for event, skipping", tableId.toString(), log);
-        continue;
-      }
-      if (!tableMetadata) {
-        debug("no table metadata found for event, skipping", tableId.toString(), log);
-        continue;
-      }
+    const tables = Array.from(new Set(block.logs.map((log) => log.args.table))).map((tableHex) =>
+      TableId.fromHex(tableHex)
+    );
+    // TODO: combine these once we refactor table registration
+    const tableSchemas = Object.fromEntries(
+      await Promise.all(tables.map(async (table) => [table.toHex(), await getTableSchema(table)]))
+    ) as Record<Hex, StoredTableSchema>;
+    const tableMetadatas = Object.fromEntries(
+      await Promise.all(tables.map(async (table) => [table.toHex(), await getTableMetadata(table)]))
+    ) as Record<Hex, StoredTableMetadata>;
 
-      console.log("log", log, tableSchema, tableMetadata, {
-        blockNumber: block.blockNumber,
-        blockHash: block.blockHash,
-        logs: [log],
-      });
+    const operations = block.logs
+      .map((log): StorageOperation<TConfig> | undefined => {
+        const tableId = TableId.fromHex(log.args.table);
+        const tableSchema = tableSchemas[log.args.table];
+        const tableMetadata = tableMetadatas[log.args.table];
+        if (!tableSchema) {
+          debug("no table schema found for event, skipping", tableId.toString(), log);
+          return;
+        }
+        if (!tableMetadata) {
+          debug("no table metadata found for event, skipping", tableId.toString(), log);
+          return;
+        }
 
-      const keyTupleValues = decodeKeyTuple(tableSchema.schema.keySchema, log.args.key);
-      const keyTuple = Object.fromEntries(
-        keyTupleValues.map((value, i) => [tableMetadata.keyNames[i] ?? i, value])
-      ) as Key<TConfig, keyof TConfig["tables"]>;
+        const keyTupleValues = decodeKeyTuple(tableSchema.schema.keySchema, log.args.key);
+        const keyTuple = Object.fromEntries(
+          keyTupleValues.map((value, i) => [tableMetadata.keyNames[i] ?? i, value])
+        ) as Key<TConfig, keyof TConfig["tables"]>;
 
-      if (log.eventName === "StoreSetRecord" || log.eventName === "StoreEphemeralRecord") {
-        const values = decodeRecord(tableSchema.schema.valueSchema, log.args.data);
-        const record = Object.fromEntries(tableMetadata.valueNames.map((name, i) => [name, values[i]])) as Value<
-          TConfig,
-          keyof TConfig["tables"]
-        >;
-        // TODO: decide if we should handle ephemeral records separately?
-        //       they'll eventually be turned into "events", but unclear if that should translate to client storage operations
-        await setRecord({ ...tableId, keyTuple, record });
-      } else if (log.eventName === "StoreSetField") {
-        const valueName = tableMetadata.valueNames[log.args.schemaIndex] as string &
-          keyof Value<TConfig, keyof TConfig["tables"]>;
-        const value = decodeField(
-          schemaIndexToAbiType(tableSchema.schema.valueSchema, log.args.schemaIndex),
-          log.args.data
-        ) as Value<TConfig, keyof TConfig["tables"]>[typeof valueName];
-        console.log("setting field", { ...tableId, keyTuple, valueName, value });
-        await setField({ ...tableId, keyTuple, valueName, value });
-      } else if (log.eventName === "StoreDeleteRecord") {
-        await deleteRecord({ ...tableId, keyTuple });
-      } else {
-        debug("unknown store event, skipping", log);
-      }
-    }
+        if (log.eventName === "StoreSetRecord" || log.eventName === "StoreEphemeralRecord") {
+          const values = decodeRecord(tableSchema.schema.valueSchema, log.args.data);
+          const record = Object.fromEntries(tableMetadata.valueNames.map((name, i) => [name, values[i]])) as Value<
+            TConfig,
+            keyof TConfig["tables"]
+          >;
+          // TODO: decide if we should handle ephemeral records separately?
+          //       they'll eventually be turned into "events", but unclear if that should translate to client storage operations
+          return {
+            address: log.address,
+            type: "SetRecord",
+            ...tableId,
+            keyTuple,
+            record,
+          };
+        }
+
+        if (log.eventName === "StoreSetField") {
+          const valueName = tableMetadata.valueNames[log.args.schemaIndex] as string &
+            keyof Value<TConfig, keyof TConfig["tables"]>;
+          const value = decodeField(
+            schemaIndexToAbiType(tableSchema.schema.valueSchema, log.args.schemaIndex),
+            log.args.data
+          ) as Value<TConfig, keyof TConfig["tables"]>[typeof valueName];
+          console.log("setting field", { ...tableId, keyTuple, valueName, value });
+          return {
+            address: log.address,
+            type: "SetField",
+            ...tableId,
+            keyTuple,
+            valueName,
+            value,
+          };
+        }
+
+        if (log.eventName === "StoreDeleteRecord") {
+          return {
+            address: log.address,
+            type: "DeleteRecord",
+            ...tableId,
+            keyTuple,
+          };
+        }
+
+        debug("unknown store event or log, skipping", log);
+        return;
+      })
+      .filter(isDefined);
+
+    return {
+      blockNumber: block.blockNumber,
+      blockHash: block.blockHash,
+      operations,
+    };
   };
 }
